Move partition-store tests in `src`. This makes sure we just compile one test binary + lets us use internal types.
