+++ crates/partition-store/src/tests/journal_table_test/mod.rs
@@ -8,7 +8,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use crate::storage_test_environment;
+use super::storage_test_environment;
 
 use bytes::Bytes;
 use bytestring::ByteString;
