+++ crates/partition-store/src/tests/outbox_table_test/mod.rs
@@ -8,8 +8,9 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use crate::mock_random_service_invocation;
-use restate_partition_store::PartitionStore;
+use super::mock_random_service_invocation;
+
+use crate::PartitionStore;
 use restate_storage_api::outbox_table::{OutboxMessage, OutboxTable};
 use restate_storage_api::Transaction;
