+++ crates/partition-store/src/tests/virtual_object_status_table_test/mod.rs
@@ -8,7 +8,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use restate_partition_store::PartitionStore;
+use crate::PartitionStore;
 use restate_storage_api::service_status_table::{VirtualObjectStatus, VirtualObjectStatusTable};
 use restate_types::identifiers::{InvocationId, InvocationUuid, ServiceId};
